+++ src/tests/commands/create/core.test.ts
